gestor_servicio_tecnico/requirements.txt
flet
psycopg2
python-dotenv
